refactor(carRacing) - Rename package and remove before version
