--- src/main/kotlin/wk1_step4/Car.kt
@@ -1,3 +1,0 @@
-package wk1_step4
-
-class Car(val name: String, var position: Int)
