--- src/main/kotlin/wk1_step4/WinnerCarRacingException.kt
@@ -1,3 +1,0 @@
-package wk1_step4
-
-class WinnerCarRacingException(message: String) : Exception(message)
