+++ src/test/kotlin/carracing/Test.kt
@@ -1,0 +1,3 @@
+package carracing
+
+class Test
