+++ src/test/kotlin/person/Test.kt
@@ -1,10 +1,10 @@
-package wk1_step1
+package person
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
 
-class PersonTest {
+class Test {
 
     @Test
     fun constructor() {
